Moved useCurrencyConverter custom hook to hooks folder
useCurrencyConverter is now in hooks folder to help clarify the role of different parts of the codebase
